testing to see gui's ability to display information
It seems that the Qt GUI is unable to display information at high
quantities and frequencies. I have tested a simple "for loop" with the
number changing from 1 to 10,000 and displaying this number to the user
as it changes, which works, but the application crashes in the middle
when I set the number to change from 1 to 100,000 instead. I am
currently unsure how to fix this problem.