Added correction to data writing logic
First, Count is modified based on channels which means it is usually a
few data points larger than what is being written.  This means that we
were losing 1 data point per channel every time the data was sent to the
binary file.

Secondly, if a file was stopped the program will now flush the remaining
data points in the memory buffer instead of losing half the buffer.

For the future, we will need to add an option to (instead of flushing
the memory buffer) adjust so that the desired user duration is used if
that option is selected.